# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: 8 Degree Luxury Real Estate API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [ongoing, completed, upcoming]
        - name: area
          in: query
          schema:
            type: string
        - name: property_type
          in: query
          schema:
            type: string
        - name: bedrooms
          in: query
          schema:
            type: integer
        - name: price_min
          in: query
          schema:
            type: integer
        - name: price_max
          in: query
          schema:
            type: integer
        - name: featured
          in: query
          schema:
            type: boolean
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: object
                properties:
                  projects:
                    type: array
                    items:
                      $ref: "#/components/schemas/Project"
                  total:
                    type: integer
                required: [projects, total]

    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectBody"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
...
  /projects/{slug}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectBody"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"

    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /projects/{slug}/units:
    get:
      operationId: getProjectUnits
      tags: [units]
      summary: Get units for a project
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project units
          content:
            application/json:
              schema:
                type: object
                properties:
                  units:
                    type: array
                    items:
                      $ref: "#/components/schemas/Unit"
                required: [units]
  /units:
    get:
      operationId: listUnits
      tags: [units]
      summary: List all units
      parameters:
        - name: project_id
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [available, reserved, sold]
      responses:
        "200":
          description: List of units
          content:
            application/json:
              schema:
                type: object
                properties:
                  units:
                    type: array
                    items:
                      $ref: "#/components/schemas/Unit"
                required: [units]

    post:
      operationId: createUnit
      tags: [units]
      summary: Create a unit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUnitBody"
      responses:
        "201":
          description: Created unit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Unit"
  /units/{id}:
    put:
      operationId: updateUnit
      tags: [units]
      summary: Update a unit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUnitBody"
      responses:
        "200":
          description: Updated unit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Unit"

    delete:
      operationId: deleteUnit
      tags: [units]
      summary: Delete a unit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required: [success]
    Project:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        status:
          type: string
          enum: [ongoing, completed, upcoming]
        area:
          type: string
        propertyType:
          type: string
        bedroomsMin:
          type: integer
        bedroomsMax:
          type: integer
        priceFrom:
          type: integer
        currency:
          type: string
        completionDate:
          type: string
          nullable: true
        shortDescription:
          type: string
        heroImageUrl:
          type: string
          nullable: true
        featured:
          type: boolean
        unitsLeft:
          type: integer
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - title
        - slug
        - status
        - area
        - propertyType
        - bedroomsMin
        - bedroomsMax
        - priceFrom
        - currency
        - shortDescription
        - featured
        - createdAt
    ProjectDetail:
      allOf:
        - $ref: "#/components/schemas/Project"
        - type: object
          properties:
            fullDescription:
              type: string
            amenities:
              type: array
              items:
                type: string
            investmentHighlights:
              type: array
              items:
                type: string
            brochureUrl:
              type: string
              nullable: true
            images:
              type: array
              items:
                $ref: "#/components/schemas/ProjectImage"
            units:
              type: array
              items:
                $ref: "#/components/schemas/Unit"
          required: [fullDescription, amenities, investmentHighlights, images, units]
    ProjectImage:
      type: object
      properties:
        id:
          type: integer
        imageUrl:
          type: string
        caption:
          type: string
          nullable: true
        sortOrder:
          type: integer
      required: [id, imageUrl, sortOrder]
    CreateProjectBody:
      type: object
      properties:
        title:
          type: string
        slug:
          type: string
        status:
          type: string
          enum: [ongoing, completed, upcoming]
        area:
          type: string
        propertyType:
          type: string
        bedroomsMin:
          type: integer
        bedroomsMax:
          type: integer
        priceFrom:
          type: integer
        currency:
          type: string
        completionDate:
          type: string
          nullable: true
        shortDescription:
          type: string
        fullDescription:
          type: string
        amenities:
          type: array
          items:
            type: string
        investmentHighlights:
          type: array
          items:
            type: string
        brochureUrl:
          type: string
          nullable: true
        heroImageUrl:
          type: string
          nullable: true
        featured:
          type: boolean
        unitsLeft:
          type: integer
          nullable: true
      required:
        - title
        - slug
        - status
        - area
        - propertyType
        - bedroomsMin
        - bedroomsMax
        - priceFrom
        - currency
        - shortDescription
        - fullDescription
        - amenities
        - investmentHighlights
        - featured
    Unit:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        unitName:
          type: string
        bedrooms:
          type: integer
        bathrooms:
          type: integer
        buildSize:
          type: integer
          nullable: true
        landSize:
          type: integer
          nullable: true
        price:
          type: integer
        currency:
          type: string
        status:
          type: string
          enum: [available, reserved, sold]
        floorplanUrl:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - projectId
        - unitName
        - bedrooms
        - bathrooms
        - price
        - currency
        - status
        - createdAt
    CreateUnitBody:
      type: object
      properties:
        projectId:
          type: integer
        unitName:
          type: string
        bedrooms:
          type: integer
        bathrooms:
          type: integer
        buildSize:
          type: integer
          nullable: true
        landSize:
          type: integer
          nullable: true
        price:
          type: integer
        currency:
          type: string
        status:
          type: string
          enum: [available, reserved, sold]
        floorplanUrl:
          type: string
          nullable: true
      required:
        - projectId
        - unitName
        - bedrooms
        - bathrooms
        - price
        - currency
        - status
...